add: cache works with movie types, not perfect

+++ app/src/main/java/com/projectapp/moviesapp/data/datasource/local/LocalDataSource.kt
@@ -16,8 +16,6 @@ interface LocalDataSource {
 
     suspend fun geAllGenresFromDb(): List<Genre>
 
-    suspend fun getGenreById(id: Int): Genre
-
+    suspend fun getGenreById(id: Long): Genre
 
 }
-
